[CastIt] Moved some main vm´s code to another partial class. Added some logic to handle changes on folders
